# linux: zvol: avoid heap allocation for zvol_request_sync=1