@problame 2021-03-01
with patch:

```
for i in 4; do fio  --allow_file_create=0 --filename_format='/dev/zvol/dut/zv$jobnum' --rw=randwrite --bs=4k --name=foo --time_based --size=500M --group_reporting=1 --sync=1 --runtime=30s  --numjobs=$i --direct=1; done
foo: (g=0): rw=randwrite, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=psync, iodepth=1
...
fio-3.23-28-g7064
Starting 4 processes
foo: (groupid=0, jobs=4): err= 0: pid=26116: Mon Mar  1 15:42:24 2021
  write: IOPS=194k, BW=758MiB/s (795MB/s)(22.2GiB/30001msec); 0 zone resets
    clat (usec): min=5, max=28048, avg=19.78, stdev=38.25
     lat (usec): min=5, max=28048, avg=19.88, stdev=38.30
    clat percentiles (usec):
     |  1.00th=[    9],  5.00th=[   10], 10.00th=[   12], 20.00th=[   15],
     | 30.00th=[   17], 40.00th=[   19], 50.00th=[   21], 60.00th=[   23],
     | 70.00th=[   24], 80.00th=[   25], 90.00th=[   26], 95.00th=[   28],
     | 99.00th=[   33], 99.50th=[   35], 99.90th=[   45], 99.95th=[   66],
     | 99.99th=[  251]
   bw (  KiB/s): min=641040, max=877257, per=100.00%, avg=777737.97, stdev=11240.09, samples=236
   iops        : min=160260, max=219314, avg=194434.37, stdev=2810.03, samples=236
  lat (usec)   : 10=8.37%, 20=40.47%, 50=51.08%, 100=0.05%, 250=0.02%
  lat (usec)   : 500=0.01%, 750=0.01%, 1000=0.01%
  lat (msec)   : 2=0.01%, 4=0.01%, 10=0.01%, 20=0.01%, 50=0.01%
  cpu          : usr=5.65%, sys=93.06%, ctx=7624, majf=0, minf=50
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=0,5822483,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=1

Run status group 0 (all jobs):
  WRITE: bw=758MiB/s (795MB/s), 758MiB/s-758MiB/s (795MB/s-795MB/s), io=22.2GiB (23.8GB), run=30001-30001msec
```

before:

```
for i in 4; do fio  --allow_file_create=0 --filename_format='/dev/zvol/dut/zv$jobnum' --rw=randwrite --bs=4k --name=foo --time_based --size=500M --group_reporting=1 --sync=1 --runtime=30s  --numjobs=$i --direct=1; done
foo: (g=0): rw=randwrite, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=psync, iodepth=1
...
fio-3.23-28-g7064
Starting 4 processes
foo: (groupid=0, jobs=4): err= 0: pid=15751: Mon Mar  1 15:40:07 2021
  write: IOPS=193k, BW=753MiB/s (790MB/s)(22.1GiB/30001msec); 0 zone resets
    clat (usec): min=5, max=24052, avg=19.94, stdev=39.98
     lat (usec): min=5, max=24052, avg=20.05, stdev=39.98
    clat percentiles (usec):
     |  1.00th=[    9],  5.00th=[   10], 10.00th=[   12], 20.00th=[   15],
     | 30.00th=[   17], 40.00th=[   19], 50.00th=[   21], 60.00th=[   23],
     | 70.00th=[   24], 80.00th=[   25], 90.00th=[   26], 95.00th=[   28],
     | 99.00th=[   34], 99.50th=[   36], 99.90th=[   46], 99.95th=[   69],
     | 99.99th=[  269]
   bw (  KiB/s): min=632793, max=859472, per=100.00%, avg=773485.20, stdev=11229.87, samples=236
   iops        : min=158198, max=214868, avg=193371.19, stdev=2807.46, samples=236
  lat (usec)   : 10=8.14%, 20=40.35%, 50=51.43%, 100=0.05%, 250=0.02%
  lat (usec)   : 500=0.01%, 750=0.01%, 1000=0.01%
  lat (msec)   : 2=0.01%, 4=0.01%, 10=0.01%, 20=0.01%, 50=0.01%
  cpu          : usr=5.50%, sys=93.23%, ctx=7404, majf=0, minf=48
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=0,5783709,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=1

Run status group 0 (all jobs):
  WRITE: bw=753MiB/s (790MB/s), 753MiB/s-753MiB/s (790MB/s-790MB/s), io=22.1GiB (23.7GB), run=30001-30001msec

```