@problame 2021-03-01
Single-Threaded

with patch
```

for i in 1; do fio  --allow_file_create=0 --filename_format='/dev/zvol/dut/zv$jobnum' --rw=randwrite --bs=4k --name=foo --time_based --size=500M --group_reporting=1 --sync=1 --runtime=30s  --numjobs=$i --direct=1; done                                                                                                                               
foo: (g=0): rw=randwrite, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=psync, iodepth=1                                                                                     
fio-3.23-28-g7064
foo: (groupid=0, jobs=1): err= 0: pid=7134: Mon Mar  1 15:45:00 2021
  write: IOPS=125k, BW=489MiB/s (513MB/s)(14.3GiB/30001msec); 0 zone resets
    clat (usec): min=5, max=9010, avg= 7.60, stdev= 5.27
     lat (usec): min=5, max=9011, avg= 7.66, stdev= 5.27
    clat percentiles (nsec):
     |  1.00th=[ 6176],  5.00th=[ 6560], 10.00th=[ 6624], 20.00th=[ 6688],
     | 30.00th=[ 6816], 40.00th=[ 6880], 50.00th=[ 6944], 60.00th=[ 7008],
     | 70.00th=[ 7136], 80.00th=[ 7264], 90.00th=[ 8640], 95.00th=[12480],
     | 99.00th=[19584], 99.50th=[22144], 99.90th=[26752], 99.95th=[28800],
     | 99.99th=[33536]
   bw (  KiB/s): min=278480, max=552008, per=100.00%, avg=500819.12, stdev=81723.38, samples=59                                                                                               
   iops        : min=69620, max=138002, avg=125204.81, stdev=20430.87, samples=59
  lat (usec)   : 10=91.27%, 20=7.79%, 50=0.94%, 100=0.01%, 250=0.01%
  lat (usec)   : 750=0.01%
  lat (msec)   : 10=0.01%
  cpu          : usr=6.89%, sys=93.06%, ctx=68, majf=0, minf=11
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=0,3755868,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=1

Run status group 0 (all jobs):
  WRITE: bw=489MiB/s (513MB/s), 489MiB/s-489MiB/s (513MB/s-513MB/s), io=14.3GiB (15.4GB), run=30001-30001msec  
```

before

```
for i in 1; do fio  --allow_file_create=0 --filename_format='/dev/zvol/dut/zv$jobnum' --rw=randwrite --bs=4k --name=foo --time_based --size=500M --group_reporting=1 --sync=1 --runtime=30s  --numjobs=$i --direct=1; done
foo: (g=0): rw=randwrite, bs=(R) 4096B-4096B, (W) 4096B-4096B, (T) 4096B-4096B, ioengine=psync, iodepth=1                                                                                     
fio-3.23-28-g7064
foo: (groupid=0, jobs=1): err= 0: pid=23381: Mon Mar  1 15:46:32 2021
  write: IOPS=121k, BW=472MiB/s (495MB/s)(13.8GiB/30001msec); 0 zone resets
    clat (usec): min=5, max=6332, avg= 7.88, stdev= 4.30
     lat (usec): min=5, max=6333, avg= 7.95, stdev= 4.31
    clat percentiles (nsec):
     |  1.00th=[ 6368],  5.00th=[ 6752], 10.00th=[ 6816], 20.00th=[ 6880],
     | 30.00th=[ 7008], 40.00th=[ 7072], 50.00th=[ 7136], 60.00th=[ 7200],
     | 70.00th=[ 7328], 80.00th=[ 7456], 90.00th=[ 9024], 95.00th=[13120],
     | 99.00th=[21120], 99.50th=[23424], 99.90th=[27520], 99.95th=[29568],
     | 99.99th=[35584]
   bw (  KiB/s): min=243208, max=537360, per=100.00%, avg=482977.76, stdev=81357.45, samples=59                                                                                               
   iops        : min=60802, max=134340, avg=120744.37, stdev=20339.27, samples=59
  lat (usec)   : 10=90.88%, 20=7.77%, 50=1.34%, 100=0.01%, 250=0.01%
  lat (usec)   : 500=0.01%, 1000=0.01%
  lat (msec)   : 10=0.01%
  cpu          : usr=7.15%, sys=92.80%, ctx=87, majf=0, minf=11
  IO depths    : 1=100.0%, 2=0.0%, 4=0.0%, 8=0.0%, 16=0.0%, 32=0.0%, >=64=0.0%
     submit    : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     complete  : 0=0.0%, 4=100.0%, 8=0.0%, 16=0.0%, 32=0.0%, 64=0.0%, >=64=0.0%
     issued rwts: total=0,3622541,0,0 short=0,0,0,0 dropped=0,0,0,0
     latency   : target=0, window=0, percentile=100.00%, depth=1

Run status group 0 (all jobs):
  WRITE: bw=472MiB/s (495MB/s), 472MiB/s-472MiB/s (495MB/s-495MB/s), io=13.8GiB (14.8GB), run=30001-30001msec
```